Allow for a gradual migration from ShadowUIModeManager to ShadowUiModeManager
There are many public APIs in ShadowUIModeManager, so it makes sense to
allow for a gradual migration to ShadowUiModeManager. Update
ShadowUiModeManager to be a base class with abstract public APIs. This
means teams can migrate from ShadowUIModeManager to ShadowUiModeManager
in 4.15, and we can remove ShadowUIModeManager in 4.16.
